refactor: split metrics rendering modules
